Fix some caching issues around child card sets.
We need to listen to the row_created event, since we can create child card
sets directly, and the caches need to be updated to reflect that.

When showing child card sets, we also need to reload the model when
we rename card sets, otherwise the model state gets confused.