Add explicit connection cleanup
Python 3.11 tweaked how the sqlite module worked, which
changed the implicit cleanup behavior, which can cause
problems on windows when temporary files are involved.
See the discussion in https://github.com/python/cpython/issues/97641

The cursor juggling code in the database upgrade test implicitly
relied on this, so we need to tweak the logic so we explicitly
close the connection so the tests pass on Windows again.